requirements.txt
asgiref==3.4.1
Django==3.2.6
pytz==2021.1
sqlparse==0.4.1

djangorestframework~=3.12.4
pillow=8.3.1